Revert "winegstreamer: Replace source pad interface with GstAppSrc.".
This reverts commit 1aa359a100bae859b278007e8bf90673eebd7db0.

appsrc suffers from a rather problematic race condition surrounding flushes [1].
Essentially, it's possible for a flushing seek to begin and end between
wg_parser_get_next_read_request and wg_parser_push_data. The race condition is
not easy to fix, and in light of it it's not clear if we want to use appsrc.

[1] https://www.winehq.org/pipermail/wine-devel/2021-September/196043.html

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=51774
Signed-off-by: Zebediah Figura <zfigura@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>