ws2_32/tests: Continue sending remaining data on short write in test_write_events.
Today, we assert that a short write clears the FD_WRITE event bit;
however, short writes can never happen on Windows in the first place.
We're testing for a property of a socket behaviour that does not exist
on Windows, but currently happens to be exhibited by Wine.

Ignore short writes, and continue sending until it fails with
EWOULDBLOCK.  This way, the test won't care whether or not a short write
clears FD_WRITE.  This allows us some flexibility in implementation of
send().

Signed-off-by: Jinoh Kang <jinoh.kang.kr@gmail.com>
Signed-off-by: Zebediah Figura <zfigura@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>