urlmon: Add basic implementation of MapBrowserEmulationModeToUserAgent.
This is necessary for mshtml's navigator userAgent to work properly, since
some apps expect it.

The second argument seems to be output pointer that gets allocated (by
CoTaskMemAlloc, or something compatible with it), but the first argument
seems to be a pointer to an unknown struct, which makes it almost impossible
to guess reliably what it does by just inspecting its behavior...

Thankfully, it seems only the first DWORD field is necessary for
this. Changing it returns the user agent for that given mode, if the rest
is zeros.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>
Signed-off-by: Jacek Caban <jacek@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>