server: Don't set error in find_object_index if the object is not found.
find_object_index() sets the status code to STATUS_NO_MORE_ENTRIES to
indicate that the given index does not exist in the list of the
directory entries, in addition to returning NULL.

STATUS_NO_MORE_ENTRIES only makes sense in the context of the function's
only caller, req_get_directory_entry(), which is used to implement the
NtQueryDirectoryObject() system call, specifically the single object
case.  Otherwise, the choice of status code is unintuitive.

Remove the set_error() call in find_object_index(), and move the
responsibility of setting the status code to the caller.