ntdll: Fail to allocate large blocks if heap isn't growable.
This tweaks the test sizes so that it passes on Wine too. The created
heap should be large enough to hold the requested alloc size, but the
allocation still fails.

Windows seems to have a specific lower threshold for large blocks with
non growable heaps, but it's probably not very useful to match it.

Signed-off-by: Rémi Bernon <rbernon@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>