wineusb.sys: Expose composite device interfaces.
On Windows this is the job of usbccgp.sys, a separate driver that layers on top
of the base USB driver. While we could take this approach as well, implementing
usbccgp is a lot more work, whereas interface support is effectively built into
libusb.

This patch helps us get closer to installing and running the Hauppauge cx231xx
drivers.

Signed-off-by: Zebediah Figura <zfigura@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>