uiautomationcore/tests: Skip UiaProviderFromIAccessible tests on Win10v1507.
In Windows 10 version 1507, UiaProviderFromIAccessible will try to check
the IAccessible passed in against the root IAccessible of the HWND
associated with the passed in IAccessible. In all other Windows
versions, this check is done on a call to get_HostRawElementProvider, or
the first time navigating towards a parent or sibling. Rather than work
around this, just skip the tests if this happens.

Signed-off-by: Connor McAdams <cmcadams@codeweavers.com>