maintainers: Add a GStreamer section.
Assign formal maintainership of the whole DLL to myself.

Although there are parts of winegstreamer that fall under the purview of Media
Foundation, it is not really possible to separate these logically from the parts
which interact with the GStreamer backend. Better to have a dedicated maintainer
for the whole DLL, and assign patches to the Media Foundation maintainer as
necessary.

Signed-off-by: Zebediah Figura <zfigura@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>