winepulse: Try to obtain shorter device names if it exceeds 62 chars.
If the device desc is longer than the limit, we try to obtain product name
or alsa name (if product name does not fit in the limit), as otherwise
there is no point to do it anyway. Then the profile desc is appended if it
doesn't cause us to exceed the limit.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=53010
Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>
Signed-off-by: Andrew Eikum <aeikum@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>