winegstreamer: Support zero-copy output using the allocator.
Through a custom allocator, by borrowing memory from the reading thread
and mapping it instead of the allocated memory.

We cannot use the buffer pool to share wrapped buffers, because some
decoder will hold on the acquired buffers longer than they should and
we cannot remove our memory from them as long as they keep a reference.

Swapping the memory on map should be safe.

Signed-off-by: Rémi Bernon <rbernon@codeweavers.com>