wined3d: Cap the maximum reported VRAM, not the current reported VRAM.
That is, subtract the currently used memory from the capped maximum, so that
creating textures will report that available VRAM has decreased. Drivers on
Windows seem to match this behaviour (at least NVidia and WARP), although they
vary on where the cap is set.

This fixes test failures in test_vidmem_accounting() for cards with over 4 GiB
of VRAM.

Signed-off-by: Zebediah Figura <zfigura@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>