winex11.drv: Change order register_extension for WGL_ARB_pixel_format_float.
Some processes use very naive string matching when detecting extensions,
for example, Cossacks III checks if the extension is "contained" in the
current extension is checking, causing that `WGL_ARB_pixel_format` is
contained in `WGL_ARB_pixel_format_float`, but it ends up with the flag
deactivated because they aren't the same.

This makes it consistent with the behavior on winemac.drv/opengl.c
too.

Signed-off-by: Sebastián Aedo <sebastian.aedo29@gmail.com>