mshtml: Only default to IE11 mode for Internet URL Zones when the app is IE.
For documents exposing a <!DOCTYPE> node but no specific X-UA-Compatible,
mshtml defaults to IE7 compat mode, unless the app is Internet Explorer and
is in Internet URL Zone. This checking for the `iexplore.exe` app name seems
hardcoded into mshtml; the FeatureControl registry keys do not affect this
directly, and none are set by default for iexplore.exe that would affect
this, anyway.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>