mshtml: Parse X-UA-Compatible correctly.
There can be multiple compat modes defined, separated by commas, but also
surrounded by whitespace. The highest mode in the list is picked as the
document mode, with 'edge' being the highest mode available.

It stops as soon as an invalid mode is found in the list and returns whatever
highest mode was found until then.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>